+++ phenodata/command.py
@@ -24,7 +24,7 @@ def run():
       phenodata info
       phenodata list-species --source=dwd [--format=csv]
       phenodata list-phases --source=dwd [--format=csv]
-      phenodata list-stations --source=dwd --dataset=immediate [--all] [--format=csv]
+      phenodata list-stations --source=dwd --dataset=immediate [--all] [--filter=berlin] [--format=csv]
       phenodata nearest-station --source=dwd --dataset=immediate --latitude=52.520007 --longitude=13.404954 [--format=csv]
       phenodata nearest-stations --source=dwd --dataset=immediate [--all] --latitude=52.520007 --longitude=13.404954 [--limit=10] [--format=csv]
       phenodata list-quality-levels --source=dwd [--format=csv]
@@ -140,7 +140,7 @@ def run():
     elif options['list-phases']:
         data = client.get_phases()
     elif options['list-stations']:
-        data = client.get_stations(all=options['all'])
+        data = client.get_stations(filter=options['filter'], all=options['all'])
     elif options['list-quality-levels']:
         data = client.get_quality_levels()
     elif options['list-quality-bytes']: