+++ CHANGES.rst
@@ -6,6 +6,7 @@ development
 ===========
 - Don't display multi-index column in "forecast" mode w/o "humanize" option
 - Grok species "Rübe" by applying appropriate fixups to raw data before parsing
+- Add option "--filter" to "phenodata list-stations" for filtering by arbitrary text strings
 
 2018-03-28 0.6.5
 ================